@@ -20,38 +20,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.3",
-    "@angular/cdk": "^14.1.3",
-    "@angular/common": "^14.1.3",
-    "@angular/compiler": "^14.1.3",
-    "@angular/core": "^14.1.3",
-    "@angular/forms": "^14.1.3",
-    "@angular/material": "^14.1.3",
-    "@angular/platform-browser": "^14.1.3",
-    "@angular/platform-browser-dynamic": "^14.1.3",
-    "@angular/router": "^14.1.3",
-    "@enigmatry/entry-components": "^1.14.9",
+    "@angular/animations": "^15.2.9",
+    "@angular/cdk": "^15.2.9",
+    "@angular/common": "^15.2.9",
+    "@angular/compiler": "^15.2.9",
+    "@angular/core": "^15.2.9",
+    "@angular/forms": "^15.2.9",
+    "@angular/material": "^15.2.9",
+    "@angular/platform-browser": "^15.2.9",
+    "@angular/platform-browser-dynamic": "^15.2.9",
+    "@angular/router": "^15.2.9",
     "@azure/msal-browser": "^2.37.0",
-    "@enigmatry/entry-form": "1.2.5",
-    "@enigmatry/entry-table": "1.2.5",
-    "@ngx-formly/core": "^5.12.6",
-    "@ngx-formly/material": "^5.12.6",
+    "@enigmatry/entry-components": "1.15.1-preview.2",
+    "@enigmatry/entry-form": "1.15.1-preview.2",
+    "@enigmatry/entry-table": "1.15.1-preview.2",
+    "@ngx-formly/core": "6.1.7",
+    "@ngx-formly/material": "6.1.7",
     "nswag": "^13.16.1",
-    "rxjs": "^6.6.7",
+    "rxjs": "^7.8.1",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.1.3",
-    "@angular/cli": "^14.1.3",
-    "@angular/compiler-cli": "^14.1.3",
-    "@angular/localize": "^14.1.3",
-    "@enigmatry/eslint-config": "1.2.5",
-    "@enigmatry/scss-foundation": "1.2.5",
-    "@enigmatry/stylelint-config": "1.2.5",
+    "@angular-devkit/build-angular": "^15.2.9",
+    "@angular/cli": "^15.2.9",
+    "@angular/compiler-cli": "^15.2.9",
+    "@angular/localize": "^15.2.9",
+    "@enigmatry/eslint-config": "1.15.1-preview.2",
+    "@enigmatry/scss-foundation": "1.15.1-preview.2",
+    "@enigmatry/stylelint-config": "1.15.1-preview.2",
     "@types/jasmine": "^3.8.1",
     "@types/node": "^16.4.3",
-    "eslint": "^8.22.0",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.4",
@@ -60,10 +59,8 @@
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
     "npm-audit-reporter-teamcity": "0.3.1",
-    "postcss-scss": "^4.0.6",
     "protractor": "~7.0.0",
-    "stylelint": "14.11.0",
     "ts-node": "~10.4.0",
-    "typescript": "4.7.4"
+    "typescript": "4.9.5"
   }
 }
