@@ -1,7 +1,5 @@
-/* eslint-disable max-len */
-/* eslint-disable no-secrets/no-secrets */
-/* eslint-disable id-length */
 import { UntypedFormControl } from '@angular/forms';
+import { firstValueFrom } from 'rxjs';
 import { ValidatorsService as CustomValidatorsService } from './validators.service';
 
 export { ValidatorsService as CustomValidatorsService } from './validators.service';
@@ -10,8 +8,8 @@ const productCodeIsUniqueValidator = async(control: UntypedFormControl, service:
     : Promise<{ productCodeIsUnique: boolean } | null> => {
     if (control.dirty) {
         const productId = control?.parent?.get('id')?.value;
-        const response = await service.isCodeUnique(productId, control.value).toPromise();
-        return response.isUnique ? null : { productCodeIsUnique: true };
+        const response = await firstValueFrom(service.isCodeUnique(productId, control.value));
+        return response?.isUnique ? null : { productCodeIsUnique: true };
     }
     return Promise.resolve(null);
 };
@@ -20,8 +18,8 @@ const productNameIsUniqueValidator = async(control: UntypedFormControl, service:
     : Promise<{ productNameIsUnique: boolean } | null> => {
     if (control.dirty) {
         const productId = control?.parent?.get('id')?.value;
-        const response = await service.isNameUnique(productId, control.value).toPromise();
-        return response.isUnique ? null : { productNameIsUnique: true };
+        const response = await firstValueFrom(service.isNameUnique(productId, control.value));
+        return response?.isUnique ? null : { productNameIsUnique: true };
     }
     return Promise.resolve(null);
 };
@@ -30,6 +28,7 @@ export const customValidatorsFactory = (service: CustomValidatorsService) => {
     return {
         validationMessages: [
             { name: 'productCodeIsUnique', message: $localize`:@@validators.productCodeIsUnique:Code is not unique` },
+            // eslint-disable-next-line no-secrets/no-secrets
             { name: 'productNameIsUnique', message: $localize`:@@validators.productNameIsUnique:Name is not unique` }
         ],
         validators: [
@@ -38,6 +37,7 @@ export const customValidatorsFactory = (service: CustomValidatorsService) => {
                 validation: (control: UntypedFormControl) => productCodeIsUniqueValidator(control, service)
             },
             {
+                // eslint-disable-next-line no-secrets/no-secrets
                 name: 'productNameIsUnique',
                 validation: (control: UntypedFormControl) => productNameIsUniqueValidator(control, service)
             }
