@@ -1,9 +1,9 @@
-import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { msalConfigFactory, MSAL_CONFIG } from './azure-ad-b2c/msal-config';
-import { MsalAuthService } from './azure-ad-b2c/msal-auth.service';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { AuthInterceptor } from './auth.interceptor';
+import { MsalAuthService } from './azure-ad-b2c/msal-auth.service';
+import { msalConfigFactory, MSAL_CONFIG } from './azure-ad-b2c/msal-config';
 
 @NgModule({
   declarations: [],
