@@ -1,11 +1,11 @@
-/* eslint-disable max-len */
 /* eslint-disable func-style */
-import { Configuration, LogLevel } from '@azure/msal-browser';
 import { InjectionToken } from '@angular/core';
+import { Configuration, LogLevel } from '@azure/msal-browser';
 import { environment } from 'src/environments/environment';
 
 export const MSAL_CONFIG = new InjectionToken<string>('MSAL_CONFIG');
 
+// eslint-disable-next-line camelcase
 export const extraQueryParameters = { ui_locales: 'nl' };
 
 function isIEOrEdge(): boolean {
@@ -29,6 +29,7 @@ export function msalConfigFactory(): Configuration {
   return {
     auth: {
       clientId: environment.azureAdB2C.clientId,
+      // eslint-disable-next-line max-len
       authority: `${environment.azureAdB2C.instance}/${environment.azureAdB2C.domain}/${environment.azureAdB2C.signUpSignInPolicyId}`,
       knownAuthorities: [environment.azureAdB2C.instance],
       redirectUri: `${location.protocol}//${location.host}`,
