@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { UntypedFormControl } from '@angular/forms';
+import { FormlyFieldConfig } from '@ngx-formly/core';
 import { FieldType } from '@ngx-formly/material/form-field';
 
 @Component({
@@ -7,6 +7,5 @@ import { FieldType } from '@ngx-formly/material/form-field';
   templateUrl: './formly-readonly-input.component.html',
   styleUrls: ['./formly-readonly-input.component.scss']
 })
-export class ReadonlyInputComponent extends FieldType {
-  formControl: UntypedFormControl;
+export class ReadonlyInputComponent extends FieldType<FormlyFieldConfig> {
 }
