@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { UntypedFormControl } from '@angular/forms';
+import { FormlyFieldConfig } from '@ngx-formly/core';
 import { FieldType } from '@ngx-formly/material/form-field';
 import { BehaviorSubject, isObservable } from 'rxjs';
 import { map } from 'rxjs/operators';
@@ -13,12 +13,11 @@ declare type SelectOption = {
   templateUrl: './formly-readonly-radio.component.html',
   styleUrls: ['./formly-readonly-radio.component.scss']
 })
-export class ReadonlyRadioComponent extends FieldType {
-    formControl: UntypedFormControl;
+export class ReadonlyRadioComponent extends FieldType<FormlyFieldConfig> {
     optionLabel = new BehaviorSubject<string>('');
     option$ = this.optionLabel.asObservable();
 
-    isArray = (): boolean => Array.isArray(this.to.options);
+    isArray = (): boolean => Array.isArray(this.props.options);
 
     getLabelFromArray(value: number) {
       const options = this.tryGetOptions();
@@ -32,8 +31,8 @@ export class ReadonlyRadioComponent extends FieldType {
     }
 
     getLabelFromObservable(value: number) {
-      if (isObservable(this.to.options)) {
-        this.to.options
+      if (isObservable(this.props.options)) {
+        this.props.options
           .pipe(map(options => options.map(opt => opt as SelectOption)))
           .subscribe(options => {
             const found = options.find(opt => opt.value === value);
@@ -47,8 +46,8 @@ export class ReadonlyRadioComponent extends FieldType {
     }
 
     private tryGetOptions = (): SelectOption[] | undefined => {
-      if (Array.isArray(this.to.options) && this.to.options.length > 0) {
-        return this.to.options.map(opt => opt as SelectOption);
+      if (Array.isArray(this.props.options) && this.props.options.length > 0) {
+        return this.props.options.map(opt => opt as SelectOption);
       }
       return undefined;
     };
