@@ -1,8 +1,8 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { MaterialModule } from '../material.module';
 import { ProductNameWithLinkCellComponent } from './product-name-with-link-cell/product-name-with-link-cell.component';
 import { ProductTypeCellComponent } from './product-type-cell/product-type-cell.component';
-import { MaterialModule } from '../material.module';
 
 @NgModule({
     declarations: [
